requirements.txt
streamlit==1.35.0
pandas==2.2.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
requests==2.32.3
plotly==5.22.0
python-dotenv==1.0.1
pydeck==0.9.1
